requirements.txt
Flask==3.0.2
gunicorn==21.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
requests
parsedatetime
firebase-admin
python-dotenv
threading